Moved InterMine resolvers to factory pattern.
Each resolver is now a factory function. This allows the same resolvers to be generated for multiple instances of the InterMine data source. All the factory functions are aggregated into a single factory function that's exported by the intermine resolvers module. This means developers only have to call a single factory to generate all resolvers for each of their InterMine data sources.