Added noImplicitAny to TypeScript config.
This required adding types to nearly every file. Notably, the resolvers all use a generic type for flexibility and ease of implementation, whereas the Intermine data source API has been extensively typed. The Intermine data source was also moved to the mixin design pattern so the class could still be defined in multiple files but in a manner that's compatible with TypeScript.