Introduced an ApiResponse type in the InterMine data source.
This type allows both data and metadata to be returned by a n InterMine data source API call. Subsequently, an InterMineSummaryResponse type was introduced for providing PageInfo objects in a requests response, and the existing Rersponse type was renamed IntermineDataResponse to better distinguish it from the new types. The existing InterMine data source models were updated to use these types.