Added generic locatedOn field to Location
This replaces the chromosome and supercontig fields, which were being used to disambiguate the underlying locatedOn relationship. This was acheived by updating the Location API in the intermine data source to use jsonobjects queries, instead of json queries, allowing class information about the Location intermine model to be recovered from responses. The Location resolver now uses this class information to resolve the typename of the locatedOn field.