Added LocatedOn type to restrict Location's locatedOn field
In the intermine schema, locatedOn has type BioEntity but in practice this field will only be a Chromosome or a Supercontig. The LocatedOn type is only extended by Chromosome and Supercontig so that these are the only types that appear in the Apollo exlorer when querying the Location's locatedOn field.