Added chromosomes entrypoint to API
The code supporting this entrypoint already existed but the entrypoint had been removed for some reason. As such, the only changes necessary in the supporting code were to add the appropriate resolver and fix a bug in the data-source API, i.e. the chromosome model needed to be joined with the sequence feature model that it extends.